NOTE: we need to review our indexes to make sure we are not using `createdAt` as first segment in our composite indexes

This fixes a problem when listing runs was very slow in some circumstances.

The improvements are based on sorting the index cardinality from less cardinal to more cardinal, since timestamps are not good filters they are better to stay after.

```sql
EXPLAIN ANALYZE SELECT 'id', created_at from "workflow_runs" JOIN (SELECT "workflows"."id" FROM "workflows" JOIN (SELECT "organizations"."id" FROM "organizations" WHERE "organizations"."id" = 'b9f230be-a075-4556-9891-19519be15a6a') AS "t1" ON "workflows"."organization_id" = "t1"."id" WHERE "workflows"."deleted_at" IS NULL) AS "t1" ON "workflow_runs"."workflow_id" = "t1"."id" WHERE "workflow_runs"."state" = 'expired' ORDER BY "workflow_runs"."created_at" DESC LIMIT 51;
```

before
```
Limit  (cost=0.70..11728.27 rows=51 width=40) (actual time=0.970..2307.775 rows=51 loops=1)
   ->  Nested Loop  (cost=0.70..68066.61 rows=296 width=40) (actual time=0.969..2307.772 rows=51 loops=1)
         ->  Nested Loop  (cost=0.70..68060.86 rows=296 width=24) (actual time=0.938..2307.722 rows=51 loops=1)
               ->  Index Scan using workflowrun_created_at_state on workflow_runs  (cost=0.43..65669.95 rows=7699 width=24) (actual time=0.051.
.2168.572 rows=187340 loops=1)
                     Index Cond: ((state)::text = 'expired'::text)
               ->  Index Scan using workflows_pkey on workflows  (cost=0.28..0.31 rows=1 width=32) (actual time=0.001..0.001 rows=0 loops=18734
0)
                     Index Cond: (id = workflow_runs.workflow_id)
                     Filter: ((deleted_at IS NULL) AND (organization_id = 'b9f230be-a075-4556-9891-19519be15a6a'::uuid))
                     Rows Removed by Filter: 1
         ->  Materialize  (cost=0.00..2.05 rows=1 width=16) (actual time=0.001..0.001 rows=1 loops=51)
               ->  Seq Scan on organizations  (cost=0.00..2.05 rows=1 width=16) (actual time=0.019..0.020 rows=1 loops=1)
                     Filter: (id = 'b9f230be-a075-4556-9891-19519be15a6a'::uuid)
                     Rows Removed by Filter: 83
 Planning Time: 2.259 ms
 Execution Time: 2307.853 ms
```

after

```
 Limit  (cost=1077.28..1077.41 rows=51 width=40) (actual time=23.320..23.327 rows=51 loops=1)
   ->  Sort  (cost=1077.28..1096.02 rows=7496 width=40) (actual time=23.318..23.321 rows=51 loops=1)
         Sort Key: workflow_runs.created_at DESC
         Sort Method: top-N heapsort  Memory: 28kB
         ->  Nested Loop  (cost=0.43..827.20 rows=7496 width=40) (actual time=2.246..23.181 rows=142 loops=1)
               ->  Nested Loop  (cost=0.00..37.54 rows=34 width=16) (actual time=0.059..1.226 rows=33 loops=1)
                     ->  Seq Scan on workflows  (cost=0.00..35.06 rows=34 width=32) (actual time=0.022..1.154 rows=33 loops=1)
                           Filter: ((deleted_at IS NULL) AND (organization_id = 'b9f230be-a075-4556-9891-19519be15a6a'::uuid))
                           Rows Removed by Filter: 852
                     ->  Materialize  (cost=0.00..2.05 rows=1 width=16) (actual time=0.001..0.002 rows=1 loops=33)
                           ->  Seq Scan on organizations  (cost=0.00..2.05 rows=1 width=16) (actual time=0.028..0.032 rows=1 loops=1)
                                 Filter: (id = 'b9f230be-a075-4556-9891-19519be15a6a'::uuid)
                                 Rows Removed by Filter: 83
               ->  Index Only Scan using workflowrun_workflow_id_state_created_at on workflow_runs  (cost=0.43..19.67 rows=356 width=24) (actua
l time=0.653..0.663 rows=4 loops=33)
                     Index Cond: ((workflow_id = workflows.id) AND (state = 'expired'::text))
                     Heap Fetches: 0
 Planning Time: 8.891 ms
 Execution Time: 23.564 ms
(18 rows)
```

I also tested filtering by `state`, `paginating` and a combination of `both`

Swapping `workflowrun_state_created_at` also greatly improved the performance of the expiration job.  

before
```
Limit  (cost=0.43..907.30 rows=100 width=304) (actual time=177.031..177.031 rows=0 loops=1)
   ->  Index Scan using workflowrun_created_at_state on workflow_runs  (cost=0.43..23270.77 rows=2566 width=304) (actual time=177.030..177.030
rows=0 loops=1)
         Index Cond: ((created_at <= '2024-12-09 22:59:32.7705+00'::timestamp with time zone) AND ((state)::text = 'initialized'::text))
 Planning Time: 0.961 ms
 Execution Time: 177.091 ms
```

after 
```
Limit  (cost=0.43..402.12 rows=100 width=304) (actual time=0.635..0.636 rows=0 loops=1)
   ->  Index Scan using workflowrun_state_created_at on workflow_runs  (cost=0.43..10307.75 rows=2566 width=304) (actual time=0.633..0.634 rows
=0 loops=1)
         Index Cond: (((state)::text = 'initialized'::text) AND (created_at <= '2024-12-09 22:59:32.7705+00'::timestamp with time zone))
 Planning Time: 2.944 ms
 Execution Time: 1.096 ms
```